refactor: move agent tools to plexe/tools package
